copilot-integration
add create_file tool to fileContextMcpServer alongside multi-range
edit_lines support

add coercion helpers for JSON encoded array/number args

exclude built in str_replace_editor/write_file/read_file/edit in favour
of kra-file-context MCP tools

register skillDirectories on session

add reasoning effort picker when selecting a Copilot model

show billing multiplier in model selection list

add session diff history keybinding (<leader>s) in Lua UI

guard cleanup against double invocation, handle SIGINT/SIGTERM for graceful shutdown

update md docs

copilot-integration - compaction thresholds
